Move test to tests/
